requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
celery==5.4.0
redis==5.0.8
httpx==0.27.2
python-dotenv==1.0.1
GitPython==3.1.43
